Return start/end profiling time in microseconds instead of milliseconds
The start and end time are now measured in microseconds and the type is int64_t. This way it seems more natural as we are going to support submilisecond sampling rate soon. Also it fixes cctest/test-cpu-profiler/ProfileStartEndTime test failure caused by comparison between long double and double.

TEST=cctest/test-cpu-profiler/ProfileStartEndTime
BUG=v8:2824
R=alph@chromium.org, bmeurer@chromium.org

Review URL: https://codereview.chromium.org/22172002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@16049 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ include/v8-profiler.h
@@ -149,16 +149,16 @@ class V8EXPORT CpuProfile {
   const CpuProfileNode* GetSample(int index) const;
 
   /**
-    * Returns time when the profile recording started (in milliseconds
+    * Returns time when the profile recording started (in microseconds
     * since the Epoch).
     */
-  double GetStartTime() const;
+  int64_t GetStartTime() const;
 
   /**
-    * Returns time when the profile recording was stopped (in milliseconds
+    * Returns time when the profile recording was stopped (in microseconds
     * since the Epoch).
     */
-  double GetEndTime() const;
+  int64_t GetEndTime() const;
 
   /**
    * Deletes the profile and removes it from CpuProfiler's list.
